renaming csproj and changing folder name which contains it
