+++ Azure.Functions.Cli.sln
@@ -21,7 +21,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Functions.Cli.Abstrac
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CoreToolsHost", "src\CoreToolsHost\CoreToolsHost.csproj", "{0333D5B6-B628-4605-A51E-D0AEE4C3F1FC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Functions.Cli.TestFramework", "test\Cli\Func.TestFramework\Azure.Functions.Cli.TestFramework.csproj", "{40C7D05D-93E9-40E9-AC37-2B36C5854477}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Functions.Cli.Test.Framework", "test\Cli\Test.Framework\Azure.Functions.Cli.Test.Framework.csproj", "{50E27986-E943-4A5D-8771-AE97C531B077}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -53,10 +53,10 @@ Global
 		{0333D5B6-B628-4605-A51E-D0AEE4C3F1FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0333D5B6-B628-4605-A51E-D0AEE4C3F1FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0333D5B6-B628-4605-A51E-D0AEE4C3F1FC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{40C7D05D-93E9-40E9-AC37-2B36C5854477}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{40C7D05D-93E9-40E9-AC37-2B36C5854477}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{40C7D05D-93E9-40E9-AC37-2B36C5854477}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{40C7D05D-93E9-40E9-AC37-2B36C5854477}.Release|Any CPU.Build.0 = Release|Any CPU
+		{50E27986-E943-4A5D-8771-AE97C531B077}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{50E27986-E943-4A5D-8771-AE97C531B077}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{50E27986-E943-4A5D-8771-AE97C531B077}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{50E27986-E943-4A5D-8771-AE97C531B077}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -69,7 +69,7 @@ Global
 		{78231B55-D243-46F1-9C7F-7831B40ED2D8} = {154FDAF2-0E86-450E-BE57-4E3D410B0FAC}
 		{BC78165E-CE5B-4303-BB8E-BC172E5B86E0} = {154FDAF2-0E86-450E-BE57-4E3D410B0FAC}
 		{0333D5B6-B628-4605-A51E-D0AEE4C3F1FC} = {5F51C958-39C0-4E0C-9165-71D0BCE647BC}
-		{40C7D05D-93E9-40E9-AC37-2B36C5854477} = {6EE1D011-2334-44F2-9D41-608B969DAE6D}
+		{50E27986-E943-4A5D-8771-AE97C531B077} = {6EE1D011-2334-44F2-9D41-608B969DAE6D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {FA1E01D6-A57B-4061-A333-EDC511D283C0}
