--- test/Cli/Func.TestFramework/Func.TestFramework.csproj
@@ -1,22 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectCapability Remove="TestContainer" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Azure.Storage.Queues" />
-    <PackageReference Include="xunit" />
-    <PackageReference Include="FluentAssertions"/>
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="$(RepoSrcRoot)Cli\Abstractions\Azure.Functions.Cli.Abstractions.csproj" />
-  </ItemGroup>
-
-</Project>
