+++ test/Cli/Test.Framework/Assertions/CommandResultExtensions.cs
@@ -4,7 +4,7 @@
 // Copied from: https://github.com/dotnet/sdk/blob/e793aa4709d28cd783712df40413448250e26fea/test/Microsoft.NET.TestFramework/Assertions/CommandResultExtensions.cs
 using Azure.Functions.Cli.Abstractions.Command;
 
-namespace Azure.Functions.Cli.TestFramework.Assertions
+namespace Azure.Functions.Cli.Test.Framework.Assertions
 {
     public static class CommandResultExtensions
     {
