+++ test/Cli/Test.Framework/Azure.Functions.Cli.Test.Framework.csproj
