+++ test/Cli/Test.Framework/Commands/FuncInitCommand.cs
@@ -3,11 +3,11 @@
 
 using Xunit.Abstractions;
 
-namespace Azure.Functions.Cli.TestFramework.Commands
+namespace Azure.Functions.Cli.Test.Framework.Commands
 {
-    public class FuncSettingsCommand(string funcPath, string testName, ITestOutputHelper log) : FuncCommand(log)
+    public class FuncInitCommand(string funcPath, string testName, ITestOutputHelper log) : FuncCommand(log)
     {
-        private readonly string _commandName = "settings";
+        private readonly string _commandName = "init";
         private readonly string _funcPath = funcPath;
         private readonly string _testName = testName;
