+++ test/Cli/Test.Framework/Commands/FuncNewCommand.cs
@@ -3,11 +3,11 @@
 
 using Xunit.Abstractions;
 
-namespace Azure.Functions.Cli.TestFramework.Commands
+namespace Azure.Functions.Cli.Test.Framework.Commands
 {
-    public class FuncStartCommand(string funcPath, string testName, ITestOutputHelper log) : FuncCommand(log)
+    public class FuncNewCommand(string funcPath, string testName, ITestOutputHelper log) : FuncCommand(log)
     {
-        private readonly string _commandName = "start";
+        private readonly string _commandName = "new";
         private readonly string _funcPath = funcPath;
         private readonly string _testName = testName;
