+++ test/Cli/Test.Framework/Helpers/FunctionAppSetupHelper.cs
@@ -1,10 +1,10 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using Azure.Functions.Cli.TestFramework.Commands;
+using Azure.Functions.Cli.Test.Framework.Commands;
 using Xunit.Abstractions;
 
-namespace Azure.Functions.Cli.TestFramework.Helpers
+namespace Azure.Functions.Cli.Test.Framework.Helpers
 {
     public static class FunctionAppSetupHelper
     {
